Refactor settings structure and improve rendering logic
Refactored aria_settings.toml for clarity, modularity, and updated key names. Updated loader functions in settings.cpp to match the new structure. Changed cell_max_capacity type to uint8_t for efficiency. Improved cell outline rendering with an ease-out function. Updated todo.md to reflect completed tasks.

+++ src/managers/food_manager/food_manager_settings.h
@@ -6,7 +6,7 @@ struct FoodManagerSettings
 {
 	inline static uint32_t cells_x;
 	inline static uint32_t cells_y;
-	inline static uint32_t cell_max_capacity;
+	inline static uint8_t cell_max_capacity;
 	inline static size_t update_freq; // food do not move that often so they dont have to be updated in the grid every frame
 
 	inline static unsigned max_food;
